+++ utils.go
@@ -39,7 +39,7 @@ var runCommand = func(c *exec.Cmd) ([]byte, error) {
 }
 
 func reportError(cfg Config, action string, err error) error {
-	log.Printf("executing [%s] failed, see error above and or below.\n", action)
+	log.Printf("executing [%s] failed, error: [%v]\n", action, err)
 
 	log.Println("checking gmig config ...")
 	fmt.Println(cfg.ToJSON())