+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Changelist of gmig releases
 
+## v1.2.0, 2018-08-22
+
+- the up command can now optionally stop after a specified migration filename.
+- initial configuration has sample environment variable.
+- better argument usage documentation.
+
+## v1.1.1, 2018-08-20
+
+- improve error report when YAML file is not valid.
+
 ## v1.1.0, 2018-08-12
 
 - support `.yml` extension for migration files.